pyproject.toml
[tool.poetry]
name = "gemini-chatbot-with-voice"
version = "0.1.0"
description = ""
authors = ["yuzuki"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
streamlit = "^1.41.1"
requests = "^2.32.3"
google-generativeai = "^0.8.3"
python-dotenv = "^1.0.1"
langchain-google-genai = "^2.0.7"
pillow = "^11.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
